third_party/blink/web_tests/fast/frames/out-of-document-iframe-has-child-frame-expected.txt

This tests that several ways of making an iframe that isn't inserted into a document tree but has a child frame will fail.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS targetFrame1.contentWindow == undefined is true
PASS targetFrame2.contentWindow == undefined is true
PASS targetFrame3.contentWindow == undefined is true
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
